refactor: 궁성 움직임을 추상화하여 장과 사와 졸의 움직임 표현 간소화
