--- src/main/java/janggi/domain/piece/SoldierMoveStrategy.java
@@ -1,34 +1,0 @@
-package janggi.domain.piece;
-
-import janggi.domain.board.BoardSnapshot;
-import janggi.domain.dynasty.Dynasty;
-import janggi.domain.position.Direction;
-import janggi.domain.position.Position;
-import java.util.ArrayList;
-import java.util.List;
-
-public class SoldierMoveStrategy implements MoveStrategy {
-
-    private static final MoveStrategy SOLDIER_MOVE_STRATEGY = new SoldierMoveStrategy();
-
-    public static MoveStrategy instance() {
-        return SOLDIER_MOVE_STRATEGY;
-    }
-
-    @Override
-    public List<Position> findPlaceablePositions(BoardSnapshot board, Position from, Dynasty dynasty) {
-        List<Position> placeablePositions = new ArrayList<>();
-        for (Direction dir : from.directions()) {
-            if (dir.equals(dynasty.front().back())) {
-                continue;
-            }
-            from.nextPositionByDirection(dir).ifPresent(to -> {
-                if (board.isPlaceable(to, dynasty)) {
-                    placeablePositions.add(to);
-                }
-            });
-        }
-        return placeablePositions;
-    }
-
-}
